mprbackend/requirements.txt
api==0.0.7
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==3.0.7
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jsonschema==3.2.0
nose==1.3.7
parso==0.6.2
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
Pygments==2.6.1
pyrsistent==0.16.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
